+++ honeycomb/hive.py
@@ -47,7 +47,7 @@ def run_lake_query(query, engine='hive', complex_join=False):
                 'honeycomb for bucket integrity reasons.'
             )
         schema, table_name = re.search(
-            r'INSERT OVERWRITE (?:TABLE )?(\w+)\.(\w+)', query,
+            r'INSERT *OVERWRITE *TABLE *(\w+)\.(\w+)', query,
             flags=re.IGNORECASE).groups()
 
         _, table_s3_path = get_table_s3_location(table_name, schema)